[//components/safe_browsing] Move ping_manager* into core/browser
This code is browser-side code, and so it belongs in
//components/safe_browsing/core/browser/.

Bug: 1216341
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2989009
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#897701}
