+++ android_webview/browser/BUILD.gn
@@ -242,7 +242,6 @@ source_set("browser") {
     "//components/safe_browsing/content/browser/triggers",
     "//components/safe_browsing/content/browser/web_ui",
     "//components/safe_browsing/content/common:interfaces",
-    "//components/safe_browsing/core:ping_manager",
     "//components/safe_browsing/core/browser",
     "//components/safe_browsing/core/browser/db:database_manager",
     "//components/safe_browsing/core/browser/db:safebrowsing_proto",
