+++ chrome/browser/safe_browsing/BUILD.gn
@@ -41,7 +41,6 @@ static_library("safe_browsing") {
     "//components/safe_browsing/content/browser:client_side_detection",
     "//components/safe_browsing/content/browser/triggers:suspicious_site_trigger",
     "//components/safe_browsing/content/browser/web_ui",
-    "//components/safe_browsing/core:ping_manager",
     "//components/safe_browsing/core/browser",
     "//components/safe_browsing/core/browser:referrer_chain_provider",
     "//components/safe_browsing/core/browser/db:database_manager",
