+++ components/safe_browsing/content/browser/BUILD.gn
@@ -53,7 +53,6 @@ source_set("browser") {
     "//components/safe_browsing/content/browser/web_ui:web_ui",
     "//components/safe_browsing/content/common",
     "//components/safe_browsing/content/common:interfaces",
-    "//components/safe_browsing/core:ping_manager",
     "//components/safe_browsing/core:verdict_cache_manager",
     "//components/safe_browsing/core/browser",
     "//components/safe_browsing/core/browser:referrer_chain_provider",
