+++ components/safe_browsing/core/browser/ping_manager.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "components/safe_browsing/core/ping_manager.h"
+#include "components/safe_browsing/core/browser/ping_manager.h"
 
 #include <utility>
