+++ components/safe_browsing/ios/browser/BUILD.gn
@@ -23,9 +23,9 @@ source_set("allow_list") {
 source_set("unit_tests") {
   testonly = true
   deps = [
-    "//components/safe_browsing/core:ping_manager_unittest",
     "//components/safe_browsing/core:verdict_cache_manager_unittest",
     "//components/safe_browsing/core/browser",
+    "//components/safe_browsing/core/browser:ping_manager_unittest",
     "//components/safe_browsing/core/browser:safe_browsing_url_checker_unittest",
     "//components/safe_browsing/core/browser/db:unit_tests_local_db",
     "//components/safe_browsing/core/browser/realtime:unit_tests",
