+++ weblayer/BUILD.gn
@@ -712,7 +712,6 @@ source_set("weblayer_lib_base") {
       "//components/safe_browsing/android:safe_browsing_api_handler",
       "//components/safe_browsing/content/browser",
       "//components/safe_browsing/content/browser:client_side_model_loader",
-      "//components/safe_browsing/core:ping_manager",
       "//components/safe_browsing/core:verdict_cache_manager",
       "//components/safe_browsing/core/browser",
       "//components/safe_browsing/core/browser:token_fetcher",
